--- a/src/aiida/storage/sqlite_temp/backend.py
+++ b/src/aiida/storage/sqlite_temp/backend.py
@@ -43,7 +43,7 @@ class SqliteTempBackend(StorageBackend):
     and destroys it when it is garbage collected.
     """
 
-    class Model(BaseModel):
+    class Model(BaseModel, defer_build=True):
         filepath: str = Field(
             title='Temporary directory',
             description='Temporary directory in which to store data for this backend.',