@danielhollas 2024-05-13
> Appreciate your effort to optimize this manually 👍

Cheers. :-) 

>  I guess the CI test we added a long time ago to catch these regressions is not doing its job.

The tests works fine, but it only prevents regressions in verdi. I've opened #6387 with some suggestions for further improvements.

I am afraid that even though I fixed the `aiida.orm` regression, it's practical impact will be small, since for any DB interaction one needs to load `aiida.storage` anyway (and one can see that the timings for `verdi code list` did not change much here). I am quite sad that even a very simple DB interaction takes a whole second (and it can pose real issues on AiiDAlab side). I'd like to understand more where the time is being spent. By some initial investigation, a lot of time is spent in sqlalchemy initialization. Seems like building the orm models is expensive, and we seem to be paying an upfront cost for all the DB backends unfortunately. I wonder if SQL supports something like `defer_build` in pydantic?